correction libc

+++ libft/ft_memcmp.c
@@ -6,7 +6,7 @@
 /*   By: tpetit <tpetit@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/16 16:56:50 by tpetit            #+#    #+#             */
-/*   Updated: 2020/11/23 09:07:39 by tpetit           ###   ########.fr       */
+/*   Updated: 2020/12/03 10:59:36 by tpetit           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,5 +25,5 @@ int	ft_memcmp(const void *s1, const void *s2, size_t n)
 	ns2 = (unsigned char *)s2;
 	while (++i < n - 1 && ns1[i] == ns2[i])
 		;
-	return (ns1[i] - ns2[i]);
+	return ((int)(ns1[i] - ns2[i]));
 }
